window dimensions calculations were returned to main window (to avoid project related problems with new files)
(cherry picked from commit f158ce85ef7cbf9bcce9c33b24f75290b5746cc8)
