+++ ProcessorHeatmap/MainWindow.xaml.cs
@@ -37,7 +37,7 @@ namespace ProcessorHeatmap
 
             _cpuUsage = new CpuUsage();
 
-            var (cols, rows) = Utils.CalculateDimensions(_cpuUsage.CoreCount);
+            var (cols, rows) = CalculateDimensions(_cpuUsage.CoreCount);
 
             _rowCount = rows;
             _colCount = cols;
@@ -144,6 +144,19 @@ namespace ProcessorHeatmap
                 menu.IsChecked = (double)menu.Header == rate;
             }
             _refreshTimer.Interval = TimeSpan.FromSeconds(rate);
+        }
+
+        public static (int, int) CalculateDimensions(int cellsCount)
+        {
+            // work out how many columns and rows we need (closest square values)
+            double idealWidthAndHeight = Math.Sqrt(cellsCount);
+            int cols = (int)Math.Truncate(idealWidthAndHeight);
+            if (cols < idealWidthAndHeight)
+            {
+                cols++;
+            }
+            int rows = (int)Math.Truncate(idealWidthAndHeight);
+            return (cols, rows);
         }
     }
 }
